requirements.txt
coverage==4.3.4
Django==1.10.5
django-crispy-forms==1.6.1
django-multiselectfield==0.1.5
wheel==0.24.0
